+++ src/main/java/subway/line/infra/LineRepository.java
@@ -1,7 +1,7 @@
 package subway.line.infra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import subway.line.Line;
+import subway.line.domain.Line;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
 }
