+++ src/main/java/subway/station/application/StationService.java
@@ -2,7 +2,7 @@ package subway.station.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.station.Station;
+import subway.station.domain.Station;
 import subway.station.presentation.StationRequest;
 import subway.station.presentation.StationResponse;
 import subway.station.exception.StationNotFoundException;
