+++ src/main/java/subway/station/infra/StationRepository.java
@@ -1,7 +1,7 @@
 package subway.station.infra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import subway.station.Station;
+import subway.station.domain.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
